\name{ForAI-package}
\alias{ForAI-package}
\alias{ForAI}
\docType{package}
\title{
Forecasting with Artificial Intelligence
}
\description{
The package "ForAI" combines machine learning methods as artificial neural networks, support vector regression with evolutionary algorithms as genetic algorithms (GA), particle swarm optimization (PSO), evolutionary strategies (ES), etc.
}
\details{
\tabular{ll}{
Package: \tab ForAI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-01\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
