\name{svres}
\alias{svres}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Support Vector Regression with Evolutionary Strategy}
\description{
  \code{svres} is a hybrid algorithm which combines support vector regression and evolutionary strategy (uncorrelated mutation with \emph{p} step sizes) to build predictive models.
}
\usage{
  svres(X.train, Y.train, X.test, PercentValid = 20, Generations = 500, InitialGamma = 0.001, ErrorFunc = MSE, Step = TRUE, StepBoruta = FALSE, SplitRandom = FALSE, Trace = TRUE, dTrace = 10, earlyStop = 50, kfold = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
       \item{X.train}{Data matrix (numeric) containing the input values (predictors) used to train
                    the model.}
       \item{Y.train}{Response vector (numeric) used to train the model.}
       \item{X.test}{Data matrix (numeric) containing the input values (predictors) used to test
                    the model.}
       \item{PercentValid}{Percentage of the data reserved for validation. Default is \code{20}\%.}
       \item{Generations}{Number of generations considered to train the model. Default is \code{500} generations.}
       \item{InitialGamma}{Initial gamma hyperparameter. Default is \code{0.001}.}
       \item{ErrorFunc}{Error function to be minimized. The default is the function \code{MSE} from package \code{qualV}        but the function could be customized.} 
       \item{Step}{Option whether to use the stepwise regression to prescreening the input variables. Default is \code{FALSE}.}
       \item{StepBoruta}{Option whether to use the Boruta selection to prescreening the input variables. Default is \code{FALSE}.}
       \item{SplitRandom}{Option whether to split the train set randomly. Default is \code{FALSE}.}
       \item{Trace}{If \code{TRUE}, information is printed during the running of \code{svres}. Default is \code{TRUE}.}
       \item{dTrace}{Frequency which the information is pritend during the running of \code{svres}. Default is \code{10}.}
       \item{earlyStop}{Stopping criterium for the evolutionary algorithm using the number of consecutive generations without improvements. Default is \code{50}.}
        \item{kfold}{if a integer value k>1 is specified, a k-fold cross validation on the training data is performed.}
}%end arguments

\details{To achieve better results, the use of a pre-processing technique (e.g. standardization of variables) is important.
     If there are multiple minima, try different values of \code{InitialGamma}.}

\value{
  \item{hypparameter}{Hyper-parameter of the best trained model.}
  \item{forecast}{A vector of predicted values generated by the best trained model.}
  \item{svmf}{An object of class "svm" containing the fitted model.}
  \item{ffTrain}{Error value of the training based on \code{ErrorFunc}. (It is \code{NULL} if \code{kfold > 1})}
  \item{ffValid}{Error value of the validation based on \code{ErrorFunc}.}
  \item{stepvars}{Variables selected by the prescreening methods when they are used.}
}

\references{Aranildo R. Lima, Alex J. Cannon, William W. Hsieh, 
Nonlinear regression in environmental sciences by support vector machines combined with evolutionary strategy, 
Computers & Geosciences, Volume 50, January 2013, Pages 136-144, ISSN 0098-3004, 
http://www.sciencedirect.com/science/article/pii/S0098300412002269

Cherkassky, V. and Ma, Y., 2004. Practical selection of SVM parameters
     and noise estimation for SVM regression. Neural Netw. 17, 113-126.       
}
\author{Aranildo R. Lima and Alex J. Cannon}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
     ## necessary libraries
     library(qualV)
     library(e1071)
     library(FNN)

      ## generating sets
      x1 <- rnorm(1000)
      x2 <- x1^2
      y  <- x1+x2

      x.fit <- cbind(x1,x2)[1:700,]
      x.test <- cbind(x1,x2)[701:1000,]
      y.fit <- y[1:700]
      y.test <- y[701:1000]

      ## running SVR-ES
      resul <- svres(x.fit, y.fit, x.test)

      ## points are the target and lines are the forecasting
      plot(y.test[1:100])
      lines(resul$forecast[1:100])
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line